+++ man-roxygen/ex-one_continent.R
@@ -1,18 +1,30 @@
 if (interactive()) {
   library(shiny)
+  library(dplyr)
+  library(Covid19)
+  library(tidyr)
+  library(scales)
+
+  require(DT)
+  #sapply(file.path("R",list.files("R")), source)
+
+  long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
   ui <- fluidPage(
-    tagList(
-      Covid19:::golem_add_external_resources(),
-      Covid19:::mod_growth_death_rate_ui("plot")
-    )
+    tabPanel("Continents",
+             tabsetPanel(
+               tabPanel("Africa",
+                        id = "tab_global",
+    Covid19:::mod_continent_ui("cont_comparison", "africa")
+               )))
   )
-  server <- function(input, output, session) {
+  server <- function(input, output) {
 
     orig_data <- reactive({
       get_timeseries_full_data() %>%
         get_timeseries_by_contagion_day_data()
     })
     pop_data = get_pop_data()
+    countries_data_map <- Covid19:::load_countries_data_map(destpath = system.file("./countries_data", package = "Covid19"))
 
     orig_data_aggregate <- reactive({
       orig_data_aggregate <- orig_data() %>%
@@ -27,8 +39,14 @@ if (interactive()) {
                new_prevalence_rate_1M_pop = round(10^6*new_confirmed/population, digits = 3))
       orig_data_aggregate
     })
+    n = 1000; w = 7
+    data_filtered <- reactive({
+      orig_data_aggregate() %>%
+        rescale_df_contagion(n = n, w = w)
+    })
 
-    callModule(Covid19:::mod_growth_death_rate_server, "plot", orig_data_aggregate)
+    callModule(Covid19:::mod_continent_server, "cont_comparison", orig_data_aggregate = orig_data_aggregate,
+               countries_data_map, n = n, w = w, pop_data, cont = "Africa", uicont = "africa")
   }
   runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
 }